requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
